AirData is now displaying on the AirSensor page.
The Database Service file method PopulateAirData was updated due to conflicts with data types for Null values that were contained in the columns designated as floats.

ObservableCollection of AirDataModel added to the view model. Created a new method called LoadAirSensorData(). This method retrieves the data and orders it by date & time - restricted to 10 results binds it to AirData.

The Xaml file was changed to incorporate a tabular layout. The column definitions of '*' provide consistency across column sizes.